ft_strrchr.c fixed

+++ ft_strrchr.c
@@ -14,7 +14,7 @@
 
 char	*ft_strrchr(const char *s, int c)
 {
-	size_t	len;
+	int	len;
 
 	len = ft_strlen(s);
 	while (len >= 0)
